đếm số lượng ok trong 1 bài viết, sửa lỗi khi click like lần 2

+++ ajaxProcess/actionUnlike.php
@@ -8,8 +8,10 @@ if (isset($_POST['idUser']) && isset($_POST['idStatu'])) {
 	} else {
 		$commentId = null;
 	}
-	$thich->delete($_POST['idUser'],$_POST['idStatu'], $commentId);
+	if ($thich->check($_POST['idUser'],$_POST['idStatu'], $commentId)) {
+		$thich->delete($_POST['idUser'],$_POST['idStatu'], $commentId);
+	}
 
-	echo "done";
+	echo $thich->countLiked($_POST['idStatu'], $commentId);
 }
  ?>
